# web/public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Jeevanti Hospitals Management System API",
    "description": "Comprehensive API documentation for the Jeevanti Hospitals Management System (HMS) built with Next.js, Prisma, MongoDB, and NextAuth.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/appointments": {
      "get": {
        "summary": "List appointments",
        "description": "Retrieves appointments matching optional filters: status, date, or doctorId.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Appointment status",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "scheduled",
                "waiting",
                "in_progress",
                "completed",
                "cancelled"
              ]
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "description": "Date string (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "doctorId",
            "in": "query",
            "required": false,
            "description": "Filter by Doctor ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointments retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Appointment"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "post": {
        "summary": "Book appointment",
        "description": "Books a new appointment and assigns a sequential token number.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment booked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "409": {
            "description": "Conflict - Time slot is already booked for this doctor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
    "/api/admissions": {
      "get": {
        "summary": "List admissions",
        "description": "Retrieves a list of inpatient admissions. Optional filter by status.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "admitted",
                "discharged"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Admission"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "post": {
        "summary": "Admit patient (IPD)",
        "description": "Creates an inpatient admission and marks bed status as occupied.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdmissionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admission"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
    "/api/admissions/{id}": {
      "get": {
        "summary": "Get admission details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionDetails"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "put": {
        "summary": "Update admission or Discharge patient",
        "description": "If body status is 'discharged', executes discharge process, sets bed status to available, adds/updates IPD invoice. Otherwise, updates admission details (doctor, ward, room, bed).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DischargeInput"
                  },
                  {
                    "$ref": "#/components/schemas/AdmissionUpdateInput"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admission updated or patient discharged successfully",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Admission"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "invoiceId": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "context": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "Patient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "uhid": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "age": {
            "type": "integer"
          },
          "address": {
            "type": "string"
          },
          "bloodGroup": {
            "type": "string"
          },
          "emergencyContact": {
            "type": "string"
          },
          "emergencyContactNumber": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Appointment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "departmentId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "time": {
            "type": "string"
          },
          "consultationType": {
            "type": "string",
            "enum": [
              "new",
              "follow_up"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "waiting",
              "in_progress",
              "completed",
              "cancelled"
            ]
          },
          "appointmentNumber": {
            "type": "string"
          },
          "tokenNumber": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          },
          "doctor": {
            "$ref": "#/components/schemas/Doctor"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          }
        }
      },
      "AppointmentInput": {
        "type": "object",
        "required": [
          "patientId",
          "doctorId",
          "departmentId",
          "date",
          "time",
          "consultationType"
        ],
        "properties": {
          "patientId": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "departmentId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "description": "YYYY-MM-DD"
          },
          "time": {
            "type": "string",
            "description": "Slot time (e.g. 10:00 AM)"
          },
          "consultationType": {
            "type": "string",
            "enum": [
              "new",
              "follow_up"
            ]
          }
        }
      },
...
      "Admission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "wardId": {
            "type": "string"
          },
          "roomId": {
            "type": "string"
          },
          "bedId": {
            "type": "string"
          },
          "admissionNumber": {
            "type": "string"
          },
          "admissionDate": {
            "type": "string",
            "format": "date-time"
          },
          "dischargeDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "admitted",
              "discharged"
            ]
          },
          "dischargeSummary": {
            "type": "string"
          },
          "finalDiagnosis": {
            "type": "string"
          },
          "followUpDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AdmissionInput": {
        "type": "object",
        "required": [
          "patientId",
          "doctorId",
          "wardId",
          "roomId",
          "bedId"
        ],
        "properties": {
          "patientId": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "wardId": {
            "type": "string"
          },
          "roomId": {
            "type": "string"
          },
          "bedId": {
            "type": "string"
          }
        }
      },
      "DischargeInput": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "discharged"
            ]
          },
          "dischargeSummary": {
            "type": "string"
          },
          "finalDiagnosis": {
            "type": "string"
          },
          "followUpDate": {
            "type": "string"
          }
        }
      },
      "AdmissionUpdateInput": {
        "type": "object",
        "properties": {
          "doctorId": {
            "type": "string"
          },
          "wardId": {
            "type": "string"
          },
          "roomId": {
            "type": "string"
          },
          "bedId": {
            "type": "string"
          }
        }
      },
      "AdmissionDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Admission"
          },
          {
            "type": "object",
            "properties": {
              "patient": {
                "$ref": "#/components/schemas/Patient"
              },
              "doctor": {
                "$ref": "#/components/schemas/Doctor"
              },
              "ward": {
                "$ref": "#/components/schemas/Ward"
              },
              "room": {
                "$ref": "#/components/schemas/Room"
              },
              "bed": {
                "$ref": "#/components/schemas/Bed"
              },
              "dailyRounds": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DailyRound"
                }
              },
              "invoices": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          }
        ]
      },
      "DailyRound": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "admissionId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "temperature": {
            "type": "number"
          },
          "bloodPressure": {
            "type": "string"
          },
          "pulse": {
            "type": "integer"
          },
          "respiratoryRate": {
            "type": "integer"
          },
          "oxygenSaturation": {
            "type": "integer"
          },
          "notes": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          }
        }
      },
...
      "Ward": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "general",
              "icu",
              "emergency",
              "private"
            ]
          },
          "floor": {
            "type": "integer"
          }
        }
      },
...
      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "wardId": {
            "type": "string"
          },
          "roomNumber": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "general",
              "private",
              "semi_private"
            ]
          },
          "bedCount": {
            "type": "integer"
          },
          "chargesPerDay": {
            "type": "number"
          }
        }
      },
...
      "Bed": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "roomId": {
            "type": "string"
          },
          "bedNumber": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "occupied",
              "maintenance"
            ]
          },
          "currentPatientId": {
            "type": "string"
          }
        }
      },
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "admissionId": {
            "type": "string"
          },
          "appointmentId": {
            "type": "string"
          },
          "invoiceNumber": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "OPD",
              "IPD"
            ]
          },
          "subtotal": {
            "type": "number"
          },
          "tax": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "partial",
              "paid",
              "cancelled"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "paidAt": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer"
                },
                "unitPrice": {
                  "type": "number"
                },
                "total": {
                  "type": "number"
                },
                "type": {
                  "type": "string"
                }
              }
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "amount": {
                  "type": "number"
                },
                "method": {
                  "type": "string"
                },
                "transactionId": {
                  "type": "string"
                },
                "paidAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "receivedBy": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "super_admin",
              "hospital_admin",
              "receptionist",
              "doctor",
              "nurse",
              "lab_technician",
              "pharmacist",
              "billing_staff"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Department": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "Doctor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "departmentId": {
            "type": "string"
          },
          "specialization": {
            "type": "string"
          },
          "qualification": {
            "type": "string"
          },
          "available": {
            "type": "boolean"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          }
        }
      },
...
    },
    "responses": {
      "400ValidationError": {
        "description": "Validation failed / Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "404NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "500InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}